requirements.txt
absl-py==2.3.0
ale-py==0.11.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chex==0.1.89
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
distrax==0.1.5
dm-env==1.6
dm-reverb==0.14.0
dm-tree==0.1.8
etils==1.5.2
Farama-Notifications==0.0.4
flatbuffers==25.2.10
flax==0.8.5
fonttools==4.58.4
fsspec==2025.5.1
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.73.0
gymnasium==1.1.1
h5py==3.14.0
humanize==4.12.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jaxlib==0.4.30
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==2.2.1
msgpack==1.1.1
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.6.4
packaging==25.0
pandas==2.3.0
pillow==11.2.1
portpicker==1.6.0
proglog==0.1.12
protobuf==4.25.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rlax==0.1.7
rsa==4.9.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
tensorstore==0.1.69
termcolor==3.1.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.23.0
